feat/1193 - renamed a folder
