+++ src/components/CohortBuilder/ParticipantsTableView/index.js
