+++ src/components/CohortBuilder/ParticipantsTableView/queries.js
